+++ Magic/CHANGELOG.md
@@ -33,6 +33,7 @@
  - Some potter wand upgrades changed keys to include `_upgrade`
    and not  conflict with survival wands (e.g staff_upgrade)
  - The inactive_icon feature will now work on non-inventory-mode wands
+ - Recipe discovery will now check for crafting permission
 
 ## Additions